all of edits and delete are complete
