+++ src/scripts/event.js
@@ -3,6 +3,7 @@ import { API } from "./api/api_manager.js";
 import { utilityFunc } from "./utility.js";
 import { registerUserForm } from "./login_register.js";
 import { newsFunc } from "./newscomp.js";
+import { RENDER } from "./render.js"
 
 let targetContainer = document.querySelector("#container");
 
@@ -21,7 +22,7 @@ const EVENT = {
                     sessionStorage.setItem("user_name", user[0].user_name)
                     sessionStorage.setItem("id", user[0].id)
                     let userID = sessionStorage.getItem("id")
-                    newsFunc.newsFromApi(userID)
+                    RENDER.newsFromApi(userID)
                 }
             });
         })
@@ -51,6 +52,7 @@ const EVENT = {
 };
 
 let newsEvents = {
+    // event listener for the edit button to put to the API
     editFormListener: function () {
         document.querySelector("#update-news-save-btn").addEventListener("click", () => {
             let newsTitleUp = document.querySelector("#news-title-edit").value
@@ -64,31 +66,52 @@ let newsEvents = {
             console.log("updateObj", updateNewsObj)
             API.updateApi("news", updateNewsObj)
                 .then(() => {
-                    newsFunc.newsFromApi(userId)
+                    RENDER.newsFromApi(userId)
                 })
         })
     },
-    // editDeleteBtnListener: function (element) {
-    //     element.addEventListener("click", () => {
-    //         // console.log("delete", "you clicked here")
-    //         if (event.target.id.startsWith("delete")) {
-    //             let id = event.target.id.split("-")[1]
-    //             console.log(id)
-    //             API.deleteFromApi("news", id)
-    //                 .then(data => {
-    //                     newsFunc.newsFromApi(userId)
+    // event listener for the delete button
+    editDeleteBtnListener: function (element, info, newsContainer, userId) {
+        element.addEventListener("click", () => {
+            // console.log("delete", "you clicked here")
+            if (event.target.id.startsWith("delete")) {
+                let id = event.target.id.split("-")[1]
+                console.log(id)
+                API.deleteFromApi("news", id)
+                    .then(data => {
+                        RENDER.newsFromApi(userId)
 
-    //                 })
-    //         }
-    //         if (event.target.id.startsWith("edit")) {
-    //             let id = event.target.id.split("edit")[1]
-    //             console.log("edit", id)
-    //             let editForm = newsFunc.editNewsForm(info)
-    //             console.log(editForm)
-    //             createEditForm(newsContainer.id, editForm)
-    //         }
-    //     })
-    // }
+                    })
+            }
+            if (event.target.id.startsWith("edit")) {
+                let id = event.target.id.split("edit")[1]
+                console.log("edit", id)
+                let editForm = newsFunc.editNewsForm(info)
+                console.log(editForm)
+                newsFunc.createEditForm(newsContainer.id, editForm)
+            }
+        })
+    },
+    //this saves a new post to the database
+    saveNewsBtn: function () {
+        document.querySelector("#save-btn").addEventListener("click", () => {
+            console.log("Save Button", "it works")
+            let newsTitle = document.querySelector("#news-title").value;
+            let newsSynopsis = document.querySelector("#news-synopsis").value;
+            let newsUrl = document.querySelector("#news-url").value;
+            let userId = sessionStorage.getItem("id")
+            let timeStamp = Date.now()
+            let newsAPISave = utilityFunc.createNewsObj(newsTitle, newsSynopsis, newsUrl, userId, timeStamp)
+            if (newsTitle === "" || newsSynopsis === "" || newsUrl === "") {
+                alert("Please fill in blank space")
+            } else {
+                API.saveToApi("news", newsAPISave)
+                    .then(newsAPI => {
+                        RENDER.newsFromApi(userId)
+                    })
+            }
+        })
+    }
 }
 
 export { EVENT, newsEvents }
